+++ src/app/page.tsx
@@ -1,5 +1,5 @@
 // Root: "/" → Reviews
-import ReviewPage from "@/components/reviews/ReviewPage";
+import { ReviewPage } from "@/components/reviews";
 
 export default function Page() {
   return <ReviewPage />;
