+++ src/app/reviews/page.tsx
@@ -1,5 +1,5 @@
 // src/app/reviews/page.tsx
-import ReviewPage from "@/components/reviews/ReviewPage";
+import { ReviewPage } from "@/components/reviews";
 
 export default function ReviewsRoute() {
   return <ReviewPage />;
