+++ src/components/reviews/index.ts
@@ -1,7 +1,9 @@
 // src/components/reviews/index.ts
+export { default as ReviewPage } from "./ReviewPage";
 export { default as ReviewsPage } from "./ReviewsPage";
 export { default as ReviewEditor } from "./ReviewEditor";
 export { default as ReviewList } from "./ReviewList";
 export { default as ReviewSummary } from "./ReviewSummary";
+export { default as RoleSelector } from "./RoleSelector";
 export { default as StatsStrip } from "./StatsStrip";
 export { default as ReviewPanel } from "./ReviewPanel";
